Started implementation of themes
